Finally found only correct strucle for cute_c2 linking, Collisions DETECTS!

+++ ARGO_Team_B/ARGO_Team_B/INCLUDE/PositionComponent.h
@@ -3,16 +3,16 @@
 class PositionComponent : public Component
 {
 public:
-	PositionComponent(int t_x, int t_y);
+	PositionComponent(float t_x, float t_y);
 	~PositionComponent();
-	int getPositionX();
-	int getPositionY();
-	void setPosition(int t_x, int t_y);
+	float getPositionX();
+	float getPositionY();
+	void setPosition(float t_x, float t_y);
 	void moveUp();
 	void moveDown();
 	void moveLeft();
 	void moveRight();
 private:
-	int m_currentX;
-	int m_currentY;
+	float m_currentX;
+	float m_currentY;
 };
